+++ .gitignore
@@ -7,3 +7,4 @@ jspm_packages
 .serverless
 
 .idea
+.vscode
