+++ package.json
@@ -20,6 +20,8 @@
     "jest": "^26.6.1"
   },
   "dependencies": {
+    "aws-sdk": "^2.792.0",
+    "csv-parser": "^2.3.3",
     "serverless-aws-documentation": "^1.1.0"
   }
 }
