--- product-service/product-service/.gitignore
@@ -1,11 +1,0 @@
-# package directories
-node_modules
-jspm_packages
-
-.env
-
-# Serverless directories
-.serverless
-
-# Webpack directories
-.webpack
