--- product-service/product-service/.vscode/launch.json
@@ -1,14 +1,0 @@
-{
-  "configurations": [
-    {
-      "name": "Lambda",
-      "type": "node",
-      "request": "launch",
-      "runtimeArgs": ["--inspect", "--debug-port=9229"],
-      "program": "${workspaceFolder}/node_modules/serverless/bin/serverless",
-      "args": ["offline"],
-      "port": 9229,
-      "console": "integratedTerminal"
-    }
-  ]
-}
