--- product-service/product-service/package.json
@@ -1,32 +1,0 @@
-{
-  "name": "product-service",
-  "version": "1.0.0",
-  "description": "Serverless webpack example using Typescript",
-  "main": "handler.js",
-  "scripts": {
-    "test": "jest"
-  },
-  "dependencies": {
-    "@serverless/event-mocks": "^1.1.1",
-    "pg": "^8.4.2",
-    "serverless-reqvalidator-plugin": "^1.0.3",
-    "source-map-support": "^0.5.10",
-    "ts-jest": "^26.4.3"
-  },
-  "devDependencies": {
-    "@types/aws-lambda": "^8.10.17",
-    "@types/node": "^10.12.18",
-    "@types/serverless": "^1.72.5",
-    "fork-ts-checker-webpack-plugin": "^3.0.1",
-    "jest": "^26.6.1",
-    "serverless-dotenv-plugin": "^3.1.0",
-    "serverless-webpack": "^5.2.0",
-    "ts-loader": "^5.3.3",
-    "ts-node": "^8.10.2",
-    "typescript": "^3.2.4",
-    "webpack": "^4.29.0",
-    "webpack-node-externals": "^1.7.2"
-  },
-  "author": "The serverless webpack authors (https://github.com/elastic-coders/serverless-webpack)",
-  "license": "MIT"
-}
