--- product-service/product-service/src/constants/headers.ts
@@ -1,6 +1,0 @@
-export const headers = {
-    'Content-Type': 'application/json',
-    'Access-Control-Allow-Origin': '*',
-    'Access-Control-Allow-Headers': 'Origin, X-Requested-With, Content-Type, Accept',
-    'Access-Control-Allow-Methods': 'POST, GET, OPTIONS, PATCH, PUT'
-};
