--- product-service/product-service/src/data/productsList.json
@@ -1,66 +1,0 @@
-[
-  {
-    "count": 20,
-    "description": "The Witcher 3: Wild Hunt",
-    "id": "7567ec4b-b10c-48c5-9345-fc73c48a80aa",
-    "price": 8,
-    "title": "The Witcher 3: Wild Hunt",
-    "imageUrl": "https://upload.wikimedia.org/wikipedia/ru/thumb/a/a2/The_Witcher_3-_Wild_Hunt_Cover.jpg/274px-The_Witcher_3-_Wild_Hunt_Cover.jpg"
-  },
-  {
-    "count": 15,
-    "description": "Cyberpunk 2077",
-    "id": "7567ec4b-b10c-48c5-9345-fc73c48a80a0",
-    "price": 30,
-    "title": "Cyberpunk 2077",
-    "imageUrl": "https://upload.wikimedia.org/wikipedia/ru/thumb/b/bb/%D0%9E%D0%B1%D0%BB%D0%BE%D0%B6%D0%BA%D0%B0_%D0%BA%D0%BE%D0%BC%D0%BF%D1%8C%D1%8E%D1%82%D0%B5%D1%80%D0%BD%D0%BE%D0%B9_%D0%B8%D0%B3%D1%80%D1%8B_Cyberpunk_2077.jpg/274px-%D0%9E%D0%B1%D0%BB%D0%BE%D0%B6%D0%BA%D0%B0_%D0%BA%D0%BE%D0%BC%D0%BF%D1%8C%D1%8E%D1%82%D0%B5%D1%80%D0%BD%D0%BE%D0%B9_%D0%B8%D0%B3%D1%80%D1%8B_Cyberpunk_2077.jpg"
-  },
-  {
-    "count": 10,
-    "description": "Grand Theft Auto V",
-    "id": "7567ec4b-b10c-48c5-9345-fc73c48a80a2",
-    "price": 20,
-    "title": "Grand Theft Auto V",
-    "imageUrl": "https://upload.wikimedia.org/wikipedia/ru/thumb/c/c8/GTAV_Official_Cover_Art.jpg/274px-GTAV_Official_Cover_Art.jpg"
-  },
-  {
-    "count": 10,
-    "description": "Red Dead Redemption 2",
-    "id": "7567ec4b-b10c-48c5-9345-fc73c48a80a1",
-    "price": 50,
-    "title": "Red Dead Redemption 2",
-    "imageUrl": "https://upload.wikimedia.org/wikipedia/ru/thumb/0/03/Red_Dead_Redemption_2_coverart.jpg/274px-Red_Dead_Redemption_2_coverart.jpg"
-  },
-  {
-    "count": 30,
-    "description": "Tom Clancy's Rainbow Six Siege",
-    "id": "7567ec4b-b10c-48c5-9345-fc73c48a80a2",
-    "price": 15,
-    "title": "Tom Clancy's Rainbow Six Siege",
-    "imageUrl": "https://upload.wikimedia.org/wikipedia/ru/thumb/d/db/Rrainbow6_PS4.jpg/274px-Rrainbow6_PS4.jpg"
-  },
-  {
-    "count": 15,
-    "description": "Mortal Kombat 11",
-    "id": "7567ec4b-b10c-48c5-9345-fc73348a80a1",
-    "price": 30,
-    "title": "Mortal Kombat 11",
-    "imageUrl": "https://upload.wikimedia.org/wikipedia/ru/thumb/4/4e/Mortal_Kombat_11.jpg/274px-Mortal_Kombat_11.jpg"
-  },
-  {
-    "count": 100,
-    "description": "Among Us",
-    "id": "7567ec4b-b10c-48c5-9445-fc73c48a80a2",
-    "price": 3,
-    "title": "Among Us",
-    "imageUrl": "https://upload.wikimedia.org/wikipedia/ru/thumb/8/84/Among_Us.png/274px-Among_Us.png"
-  },
-  {
-    "count": 10,
-    "description": "Short Product Description7",
-    "id": "Death Stranding",
-    "price": 50,
-    "title": "Death Stranding",
-    "imageUrl": "https://upload.wikimedia.org/wikipedia/ru/thumb/e/ee/Death_Stranding_Poster.jpg/258px-Death_Stranding_Poster.jpg"
-  }
-]
