--- product-service/product-service/src/models/product.model.ts
@@ -1,8 +1,0 @@
-export interface ProductModel {
-    id?: string;
-    title: string;
-    description: string;
-    price: number;
-    count: number;
-    imageUrl: string;
-}
